requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
requests==2.31.0
markupsafe==2.1.3
python-dotenv==1.0.0